Break out ButtonBaseAdapter subclasses
